beggining refactorization
