+++ Star Practices/src/se01/FileReaderClass.java
@@ -9,37 +9,11 @@ import java.util.Arrays;
 import java.util.List;
 
 public class FileReaderClass {
-    // Words that are used too often in english and take too much storage to index,
-    // and are practically useless
-    private List<String> stopWords = Arrays.asList("a", "able", "about",
-            "across", "after", "all", "almost", "also", "am", "among", "an",
-            "and", "any", "are", "as", "at", "be", "because", "been", "but",
-            "by", "can", "cannot", "could", "dear", "did", "do", "does",
-            "either", "else", "ever", "every", "for", "from", "get", "got",
-            "had", "has", "have", "he", "her", "hers", "him", "his", "how",
-            "however", "i", "if", "in", "into", "is", "it", "its", "just",
-            "least", "let", "like", "likely", "may", "me", "might", "most",
-            "must", "my", "neither", "no", "nor", "not", "of", "off", "often",
-            "on", "only", "or", "other", "our", "own", "rather", "said", "say",
-            "says", "she", "should", "since", "so", "some", "than", "that",
-            "the", "their", "them", "then", "there", "these", "they", "this",
-            "tis", "to", "too", "twas", "us", "wants", "was", "we", "were",
-            "what", "when", "where", "which", "while", "who", "whom", "why",
-            "will", "with", "would", "yet", "you", "your");
-
-    // blacklist characters to remove from a word
-    private char[] blacklist = "!$^()_+=-.;,[]{}\'\"".toCharArray();
 
     private List<Pair<String, Integer>> tokens = new ArrayList<Pair<String, Integer>>();
     private List<File> m_files;
 
-    private String trimWord(String _word){
-        // remove blacklist characters
-        String word = _word.toUpperCase();
-        for (char c : blacklist)
-            word = word.replace(c, '\0');
-        return word;
-    }
+    
 
     private void readFile(File file, int index) {
         // Reads a file and adds the tokens to the list
