+++ Star Practices/src/se01/Tokenizer.java
@@ -1,0 +1,33 @@
+package se01;
+
+
+public class Tokenizer {
+    // Words that are used too often in english and take too much storage to index,
+    // and are practically useless
+    private List<String> stopWords = Arrays.asList("a", "able", "about",
+            "across", "after", "all", "almost", "also", "am", "among", "an",
+            "and", "any", "are", "as", "at", "be", "because", "been", "but",
+            "by", "can", "cannot", "could", "dear", "did", "do", "does",
+            "either", "else", "ever", "every", "for", "from", "get", "got",
+            "had", "has", "have", "he", "her", "hers", "him", "his", "how",
+            "however", "i", "if", "in", "into", "is", "it", "its", "just",
+            "least", "let", "like", "likely", "may", "me", "might", "most",
+            "must", "my", "neither", "no", "nor", "not", "of", "off", "often",
+            "on", "only", "or", "other", "our", "own", "rather", "said", "say",
+            "says", "she", "should", "since", "so", "some", "than", "that",
+            "the", "their", "them", "then", "there", "these", "they", "this",
+            "tis", "to", "too", "twas", "us", "wants", "was", "we", "were",
+            "what", "when", "where", "which", "while", "who", "whom", "why",
+            "will", "with", "would", "yet", "you", "your");
+
+    // blacklist characters to remove from a word
+    private char[] blacklist = "!$^()_+=-.;,[]{}\'\"".toCharArray();
+
+    private String trimWord(String _word){
+        // remove blacklist characters
+        String word = _word.toUpperCase();
+        for (char c : blacklist)
+            word = word.replace(c, '\0');
+        return word;
+    }
+}
